// src/app/open-props/lib/sizes.stylex.d.ts
/**
 * Copyright (c) Meta Platforms, Inc. and affiliates.
 *
 * This source code is licensed under the MIT license found in the
 * LICENSE file in the root directory of this source tree.
 *
 *
 */

import type { VarGroup } from '@stylexjs/stylex/lib/StyleXTypes';
type TSizes = Readonly<{
  spacing000: string;
  spacing00: string;
  spacing1: string;
  spacing2: string;
  spacing3: string;
  spacing4: string;
  spacing5: string;
  spacing6: string;
  spacing7: string;
  spacing8: string;
  spacing9: string;
  spacing10: string;
  spacing11: string;
  spacing12: string;
  spacing13: string;
  spacing14: string;
  spacing15: string;
  fluid1: string;
  fluid2: string;
  fluid3: string;
  fluid4: string;
  fluid5: string;
  fluid6: string;
  fluid7: string;
  fluid8: string;
  fluid9: string;
  fluid10: string;
  content1: string;
  content2: string;
  content3: string;
  header1: string;
  header2: string;
  header3: string;
  xxs: string;
  xs: string;
  sm: string;
  md: string;
  lg: string;
  xl: string;
  xxl: string;
  relative000: string;
  relative00: string;
  relative1: string;
  relative2: string;
  relative3: string;
  relative4: string;
  relative5: string;
  relative6: string;
  relative7: string;
  relative8: string;
  relative9: string;
  relative10: string;
  relative11: string;
  relative12: string;
  relative13: string;
  relative14: string;
  relative15: string;
}>;
export declare const sizes: VarGroup<TSizes>;