// src/app/open-props/lib/borders.stylex.d.ts
/**
 * Copyright (c) Meta Platforms, Inc. and affiliates.
 *
 * This source code is licensed under the MIT license found in the
 * LICENSE file in the root directory of this source tree.
 *
 *
 */

import type { VarGroup } from '@stylexjs/stylex/lib/StyleXTypes';
type TBorders = Readonly<{
  size1: string;
  size2: string;
  size3: string;
  size4: string;
  size5: string;
  radius1: string;
  radius2: string;
  radius3: string;
  radius4: string;
  radius5: string;
  radius6: string;
  radiusRound: string;
  radiusConditional1: string;
  radiusConditional2: string;
  radiusConditional3: string;
  radiusConditional4: string;
  radiusConditional5: string;
  radiusConditional6: string;
}>;
export declare const borders: VarGroup<TBorders>;